Rename project to 'mazette'
Rename 'assets' projects to 'mazette', to give it a more... unique name.
The name itself is not too interesting, it just came to be when we were
thinking of names, and ran out of them while entering a brewery in
Brussels called, well, Mazette!

And it's not taken in PyPI, so that's cool.

Closes #1